Add governance update planner and reusable module plans
