+++ .gitignore
@@ -16,8 +16,9 @@ scripts/**/*.js
 !scripts/run-coverage.js
 !scripts/verify-wiring.js
 !scripts/run-wire-verify.js
-
 build/
 lib/
 agent-gateway/dist/
 examples/agentic/runtime-agentic.jsonl
+!scripts/v2/run-owner-plan.js
+!scripts/v2/lib/
